chore: major cleanup of iterators chore: changed from 8-byte size_t integers to 4-byte ones where the operations permit chore: cleanup and refactoring of may code modules

+++ examples/source/DESAdvectionDiffusion3D_Example.h
@@ -5,8 +5,8 @@
 
 namespace csmp {
 
-  template<size_t dim> class Model;
-  template<size_t dim> class VTK_Interface;
+  template<uint32_t dim> class Model;
+  template<uint32_t dim> class VTK_Interface;
 
 class  DESAdvectionDiffusion3D_Example : public Example {
   public:
