+++ source/includes/integration/finite_elements/IsoparametricLinearPyramid.h
@@ -9,39 +9,39 @@ class IsoparametricLinearPyramid : public FiniteElement {
 
   public:
 
-    explicit IsoparametricLinearPyramid(size_t integrationPoints=8 /* 1 or 8 */);
+    explicit IsoparametricLinearPyramid(uint32_t integrationPoints=8 /* 1 or 8 */);
     ~IsoparametricLinearPyramid();
 
     virtual double    Volume();
     virtual double    AspectRatio();
     virtual double    InnerRadius();
     virtual void        EdgeLengths( std::vector<double>& vec );
-    virtual void        CornerNodes( std::vector<size_t>& ids ) const;
-    virtual size_t      CornerNodes() const  { return 5; }
-    virtual void        MidSideNodes( std::vector<size_t>& ids ) const;
-    virtual void        CounterClockwiseNodes( std::vector<size_t>& ids ) const;
-    virtual size_t      MidSideNodes() const { return 0; }
-    virtual CSMP_FEM_TYPE  ElementTypeOfFace( size_t face ) const;
-    virtual CSMP_FEM_TYPE  ElementTypeOfSegment( size_t /* segment */ ) const { return ISOPARAMETRIC_LINEAR_BAR; };
-    virtual void        ConsecutiveNodesAtBoundary( const std::vector<size_t>& bnodes,
-                                                  std::vector<size_t>& fnids );
-    virtual void        NodesOfSegment( size_t segm_id, std::vector<size_t>& snids ) const;
-    virtual void        NodesOfFace( size_t face_id, std::vector<size_t>& fnids ) const;
-    virtual std::vector<size_t>  CornerNodesOfFace( size_t face_id ) const;
-    virtual std::vector<size_t>  NodesConnectedTo( size_t node_id ) const;
+    virtual void        CornerNodes( std::vector<uint32_t>& ids ) const;
+    virtual uint32_t      CornerNodes() const  { return 5; }
+    virtual void        MidSideNodes( std::vector<uint32_t>& ids ) const;
+    virtual void        CounterClockwiseNodes( std::vector<uint32_t>& ids ) const;
+    virtual uint32_t      MidSideNodes() const { return 0; }
+    virtual CSMP_FEM_TYPE  ElementTypeOfFace( uint32_t face ) const;
+    virtual CSMP_FEM_TYPE  ElementTypeOfSegment( uint32_t /* segment */ ) const { return ISOPARAMETRIC_LINEAR_BAR; };
+    virtual void        ConsecutiveNodesAtBoundary( const std::vector<uint32_t>& bnodes,
+                                                  std::vector<uint32_t>& fnids );
+    virtual void        NodesOfSegment( uint32_t segm_id, std::vector<uint32_t>& snids ) const;
+    virtual void        NodesOfFace( uint32_t face_id, std::vector<uint32_t>& fnids ) const;
+    virtual std::vector<uint32_t>  CornerNodesOfFace( uint32_t face_id ) const;
+    virtual std::vector<uint32_t>  NodesConnectedTo( uint32_t node_id ) const;
   
-    virtual void        UnitNormalToFace( size_t face, std::vector<double>& unrml ) const;
+    virtual void        UnitNormalToFace( uint32_t face, std::vector<double>& unrml ) const;
 
     virtual void        N(std::vector<double>& N, const std::vector<double>& xyz );
-    virtual void        N_AtIntegrationPoint( size_t ip, std::vector<double>& N );
+    virtual void        N_AtIntegrationPoint( uint32_t ip, std::vector<double>& N );
     virtual void        N_AtBaryCenter( std::vector<double>& N );
-    virtual void        JacobianAtIntegrationPoint( size_t ip );
+    virtual void        JacobianAtIntegrationPoint( uint32_t ip );
     virtual void        JacobianAt( const std::vector<double>& rst );
 
     virtual double    dN( DenseMatrix<DM_MIN>& dn, const std::vector<double>& xyz  );
     virtual void        dN( DenseMatrix<DM_MIN>& DN5 );
-    virtual double    dN_AtNode( DenseMatrix<DM_MIN>& M, size_t node );
-    virtual double    dN_AtIntegrationPoint( DenseMatrix<DM_MIN>& M, size_t gauss_point );
+    virtual double    dN_AtNode( DenseMatrix<DM_MIN>& M, uint32_t node );
+    virtual double    dN_AtIntegrationPoint( DenseMatrix<DM_MIN>& M, uint32_t gauss_point );
     virtual double    dN_AtBarycenter( DenseMatrix<DM_MIN>& M );
 
     virtual void        Nrst( double r, double s, double t, std::vector<double>& nrst ) const;
@@ -50,9 +50,9 @@ class IsoparametricLinearPyramid : public FiniteElement {
     virtual void        dNs( double r, double s, double t, std::vector<double>& dNs ) const;
     virtual void        dNt( double r, double s, double t, std::vector<double>& dNt ) const;
 
-    virtual double    WeightAtIntegrationPoint( size_t i ) const;
-    virtual void        IntegrationPoint( size_t i, std::vector<double>& xyz ) const;
-    virtual void        ExtrapolateIntegrationPointVariableToNodes( size_t nvars,
+    virtual double    WeightAtIntegrationPoint( uint32_t i ) const;
+    virtual void        IntegrationPoint( uint32_t i, std::vector<double>& xyz ) const;
+    virtual void        ExtrapolateIntegrationPointVariableToNodes( uint32_t nvars,
                                                                     const std::vector<double>& IVAR,
                                                                     std::vector<double>& NVAR ) const;
     virtual void        OutputNodeDataToVTK( const char* file_name,
@@ -67,10 +67,10 @@ class IsoparametricLinearPyramid : public FiniteElement {
     DenseMatrix<DM_MIN>     DN,NXYZ,IP;
 
     double  VolumeOfTetra(
-                size_t verticeIndex1,
-                size_t verticeIndex2,
-                size_t verticeIndex3,
-                size_t verticeIndex4
+                uint32_t verticeIndex1,
+                uint32_t verticeIndex2,
+                uint32_t verticeIndex3,
+                uint32_t verticeIndex4
                );
